+++ src/main/java/basic/product/ProductJDBCDAOClient.java
@@ -25,7 +25,7 @@ public class ProductJDBCDAOClient implements ProductDAO {
 
             while (rs.next()) {
                 Product product = new Product(rs.getInt("product_id"), rs.getString("name"),
-                        rs.getInt("product_category_id"), rs.getString("brand_name"), rs.getBigDecimal("price"),
+                        rs.getInt("product_category_id"), rs.getString("brand_name"), rs.getFloat("price"),
                         rs.getInt("age_category"));
                 productsList.add(product);
             }
@@ -48,7 +48,7 @@ public class ProductJDBCDAOClient implements ProductDAO {
             ResultSet rs = statement.executeQuery();
             while (rs.next()) {
                 product = new Product(rs.getInt("product_id"), rs.getString("name"),
-                        rs.getInt("product_category_id"), rs.getString("brand_name"), rs.getBigDecimal("price"),
+                        rs.getInt("product_category_id"), rs.getString("brand_name"), rs.getFloat("price"),
                         rs.getInt("age_category"));
             }
 
@@ -72,7 +72,7 @@ public class ProductJDBCDAOClient implements ProductDAO {
 
             while (rs.next()) {
                 Product product = new Product(rs.getInt("product_id"), rs.getString("name"),
-                        rs.getInt("product_category_id"), rs.getString("brand_name"), rs.getBigDecimal("price"),
+                        rs.getInt("product_category_id"), rs.getString("brand_name"), rs.getFloat("price"),
                         rs.getInt("age_category"));
                 productsList.add(product);
             }
